WIP AuthorizePersonaDetails
