+++ src/core/types/secret_bytes.rs
@@ -1,6 +1,5 @@
 use crate::prelude::*;
 
-#[macro_export]
 macro_rules! decl_secret_bytes {
     (
         $(
@@ -162,3 +161,5 @@ macro_rules! decl_secret_bytes {
         }
     };
 }
+
+pub(crate) use decl_secret_bytes;
