+++ src/profile/supporting_types/mod.rs
@@ -1,7 +1,17 @@
+mod account_for_display;
 mod account_or_persona;
 mod account_or_persona_uniffi_fn;
+mod accounts_for_display;
 mod accounts_or_personas;
+mod authorized_dapp_detailed;
+mod authorized_persona_detailed;
+mod detailed_authorized_personas;
 
+pub use account_for_display::*;
 pub use account_or_persona::*;
 pub use account_or_persona_uniffi_fn::*;
+pub use accounts_for_display::*;
 pub use accounts_or_personas::*;
+pub use authorized_dapp_detailed::*;
+pub use authorized_persona_detailed::*;
+pub use detailed_authorized_personas::*;
