+++ src/profile/v100/address/mod.rs
@@ -1,10 +1,14 @@
 mod access_controller_address;
+mod access_controller_address_uniffi_fn;
 mod account_address;
+mod account_address_uniffi_fn;
 mod address;
 mod address_format;
 mod component_address;
+mod component_address_uniffi_fn;
 mod entity_address;
 mod identity_address;
+mod identity_address_uniffi_fn;
 mod legacy_olympia_account_address;
 mod legacy_olympia_account_address_uniffi_fn;
 mod non_fungible_global_id;
@@ -15,19 +19,27 @@ mod non_fungible_local_id_uniffi_fn;
 mod non_fungible_resource_address;
 mod non_fungible_resource_address_uniffi_fn;
 mod package_address;
+mod package_address_uniffi_fn;
 mod pool_address;
+mod pool_address_uniffi_fn;
 mod resource_address;
 mod validator_address;
+mod validator_address_uniffi_fn;
 mod vault_address;
+mod vault_address_uniffi_fn;
 mod wrap_ret_address;
 
 pub use access_controller_address::*;
+pub use access_controller_address_uniffi_fn::*;
 pub use account_address::*;
+pub use account_address_uniffi_fn::*;
 pub use address::*;
 pub use address_format::*;
 pub use component_address::*;
+pub use component_address_uniffi_fn::*;
 pub use entity_address::*;
 pub use identity_address::*;
+pub use identity_address_uniffi_fn::*;
 pub use legacy_olympia_account_address::*;
 pub use legacy_olympia_account_address_uniffi_fn::*;
 pub use non_fungible_global_id::*;
@@ -38,8 +50,12 @@ pub use non_fungible_local_id_uniffi_fn::*;
 pub use non_fungible_resource_address::*;
 pub use non_fungible_resource_address_uniffi_fn::*;
 pub use package_address::*;
+pub use package_address_uniffi_fn::*;
 pub use pool_address::*;
+pub use pool_address_uniffi_fn::*;
 pub use resource_address::*;
 pub use validator_address::*;
+pub use validator_address_uniffi_fn::*;
 pub use vault_address::*;
+pub use vault_address_uniffi_fn::*;
 pub use wrap_ret_address::*;
