+++ src/profile/v100/address/resource_address.rs
@@ -1,5 +1,22 @@
 use crate::prelude::*;
 
+decl_ret_wrapped_address!(
+    /// Addresses identifying an asset, either fungible (Token) or non-fungible (NFT), on the Radix network, e.g.
+    /// `"resource_rdx1tknxxxxxxxxxradxrdxxxxxxxxx009923554798xxxxxxxxxradxrd"`
+    /// Being the unique identifier of the Radix Token, the Rad, on mainnet.
+    ///
+    /// There are fundamentally two different sub-types ([Scrypto's `EntityType`][entt]) of ResourceAddresses:
+    /// * GlobalFungibleResourceManager
+    /// * GlobalNonFungibleResourceManager
+    ///
+    /// Implementation wise we wrap [Radix Engine Toolkit's `CanonicalResourceAddress`][ret], and
+    /// give it UniFFI support, as a `uniffi::Record` (we also own Serde).
+    ///
+    /// [entt]: https://github.com/radixdlt/radixdlt-scrypto/blob/fc196e21aacc19c0a3dbb13f3cd313dccf4327ca/radix-engine-common/src/types/entity_type.rs
+    /// [ret]: https://github.com/radixdlt/radix-engine-toolkit/blob/34fcc3d5953f4fe131d63d4ee2c41259a087e7a5/crates/radix-engine-toolkit/src/models/canonical_address_types.rs#L236-L239
+    resource
+);
+
 impl ResourceAddress {
     pub fn is_fungible(&self) -> bool {
         self.secret_magic.is_fungible()
@@ -17,52 +34,6 @@ impl ResourceAddress {
     pub fn is_xrd_on_network(&self, id: NetworkID) -> bool {
         self == &Self::xrd_on_network(id)
     }
-}
-
-#[uniffi::export]
-pub fn resource_address_is_fungible(address: &ResourceAddress) -> bool {
-    address.is_fungible()
-}
-
-#[uniffi::export]
-pub fn resource_address_is_non_fungible(address: &ResourceAddress) -> bool {
-    address.is_non_fungible()
-}
-
-#[uniffi::export]
-pub fn new_resource_address_sample_mainnet_xrd() -> ResourceAddress {
-    ResourceAddress::sample_mainnet_xrd()
-}
-
-#[uniffi::export]
-pub fn new_resource_address_sample_mainnet_candy() -> ResourceAddress {
-    ResourceAddress::sample_mainnet_candy()
-}
-
-#[uniffi::export]
-pub fn new_resource_address_sample_mainnet_nft_gc_membership() -> ResourceAddress
-{
-    ResourceAddress::sample_mainnet_nft_gc_membership()
-}
-
-#[uniffi::export]
-pub fn new_resource_address_sample_stokenet_xrd() -> ResourceAddress {
-    ResourceAddress::sample_stokenet_xrd()
-}
-
-#[uniffi::export]
-pub fn new_resource_address_sample_stokenet_gum() -> ResourceAddress {
-    ResourceAddress::sample_stokenet_gum()
-}
-
-#[uniffi::export]
-pub fn new_resource_address_sample_stokenet_gc_tokens() -> ResourceAddress {
-    ResourceAddress::sample_stokenet_gc_tokens()
-}
-
-#[uniffi::export]
-pub fn new_resource_address_sample_stokenet_candy() -> ResourceAddress {
-    ResourceAddress::sample_stokenet_candy()
 }
 
 impl HasSampleValues for ResourceAddress {
@@ -422,69 +393,5 @@ mod tests {
     #[test]
     fn xrd_on_stokenet() {
         assert_eq!(SUT::xrd_on_network(NetworkID::Stokenet).to_string(), "resource_tdx_2_1tknxxxxxxxxxradxrdxxxxxxxxx009923554798xxxxxxxxxtfd2jc");
-    }
-}
-
-#[cfg(test)]
-mod uniffi_tests {
-    use crate::prelude::*;
-
-    use super::*;
-
-    #[allow(clippy::upper_case_acronyms)]
-    type SUT = ResourceAddress;
-
-    #[test]
-    fn new_from_bech32_get_network_id_and_address() {
-        let b32 = "resource_rdx1tknxxxxxxxxxradxrdxxxxxxxxx009923554798xxxxxxxxxradxrd";
-        let address = new_resource_address(b32.to_owned()).unwrap();
-        assert_eq!(SUT::try_from_bech32(b32).unwrap(), address);
-        assert_eq!(resource_address_network_id(&address), NetworkID::Mainnet);
-        assert_eq!(resource_address_bech32_address(&address), b32);
-    }
-
-    #[test]
-    fn is_fungible() {
-        assert!(!resource_address_is_fungible(
-            &SUT::sample_mainnet_nft_gc_membership()
-        ));
-        assert!(resource_address_is_non_fungible(
-            &SUT::sample_mainnet_nft_gc_membership()
-        ));
-
-        assert!(resource_address_is_fungible(&SUT::sample_mainnet_xrd()));
-        assert!(!resource_address_is_non_fungible(&SUT::sample_mainnet_xrd()));
-    }
-
-    #[test]
-    fn sample() {
-        assert_eq!(
-            new_resource_address_sample_mainnet_xrd(),
-            SUT::sample_mainnet_xrd()
-        );
-        assert_eq!(
-            new_resource_address_sample_mainnet_candy(),
-            SUT::sample_mainnet_candy()
-        );
-        assert_eq!(
-            new_resource_address_sample_mainnet_nft_gc_membership(),
-            SUT::sample_mainnet_nft_gc_membership()
-        );
-        assert_eq!(
-            new_resource_address_sample_stokenet_xrd(),
-            SUT::sample_stokenet_xrd()
-        );
-        assert_eq!(
-            new_resource_address_sample_stokenet_gum(),
-            SUT::sample_stokenet_gum()
-        );
-        assert_eq!(
-            new_resource_address_sample_stokenet_gc_tokens(),
-            SUT::sample_stokenet_gc_tokens()
-        );
-        assert_eq!(
-            new_resource_address_sample_stokenet_candy(),
-            SUT::sample_stokenet_candy()
-        );
     }
 }
